+++ docs/examples/django/manage.py
@@ -21,13 +21,13 @@ from opentelemetry.instrumentation.django import DjangoInstrumentor
 
 
 def main():
+    os.environ.setdefault(
+        "DJANGO_SETTINGS_MODULE", "instrumentation_example.settings"
+    )
 
     # This call is what makes the Django application be instrumented
     DjangoInstrumentor().instrument()
 
-    os.environ.setdefault(
-        "DJANGO_SETTINGS_MODULE", "instrumentation_example.settings"
-    )
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc: