+++ instrumentation/opentelemetry-instrumentation-falcon/src/opentelemetry/instrumentation/falcon/__init__.py
@@ -115,13 +115,16 @@ class _InstrumentedFalconAPI(falcon.API):
         start_time = time_ns()
 
         token = context.attach(propagators.extract(otel_wsgi.getter, env))
-        attributes = otel_wsgi.collect_request_attributes(env)
         span = self._tracer.start_span(
             otel_wsgi.get_default_span_name(env),
             kind=trace.SpanKind.SERVER,
-            attributes=attributes,
             start_time=start_time,
         )
+        if span.is_recording():
+            attributes = otel_wsgi.collect_request_attributes(env)
+            for key, value in attributes.items():
+                span.set_attribute(key, value)
+
         activation = self._tracer.use_span(span, end_on_exit=True)
         activation.__enter__()
         env[_ENVIRON_SPAN_KEY] = span
@@ -160,7 +163,7 @@ class _TraceMiddleware:
 
     def process_request(self, req, resp):
         span = req.env.get(_ENVIRON_SPAN_KEY)
-        if not span:
+        if not span or not span.is_recording():
             return
 
         attributes = extract_attributes_from_object(
@@ -171,7 +174,7 @@ class _TraceMiddleware:
 
     def process_resource(self, req, resp, resource, params):
         span = req.env.get(_ENVIRON_SPAN_KEY)
-        if not span:
+        if not span or not span.is_recording():
             return
 
         resource_name = resource.__class__.__name__
@@ -184,7 +187,7 @@ class _TraceMiddleware:
         self, req, resp, resource, req_succeeded=None
     ):  # pylint:disable=R0201
         span = req.env.get(_ENVIRON_SPAN_KEY)
-        if not span:
+        if not span or not span.is_recording():
             return
 
         status = resp.status